pyproject.toml
[project]
name = "face-identification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cython>=3.2.4",
    "insightface>=0.7.3",
    "numpy>=2.4.2",
    "onnxruntime>=1.24.2",
    "opencv-python>=4.13.0.92",
    "pgvector>=0.4.2",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.1",
    "ultralytics>=8.4.14",
]
